Added new customized icon with its different configuration sizes, organized folders structure, created XMLs for both screens
