+++ GitRepositories/app/src/main/AndroidManifest.xml
@@ -6,10 +6,10 @@
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
+        android:roundIcon="@mipmap/ic_launcher"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".modules.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
