+++ GitRepositories/app/src/main/java/com/example/gitrepositories/modules/MainActivity.kt
@@ -1,12 +1,13 @@
-package com.example.gitrepositories
+package com.example.gitrepositories.modules
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.gitrepositories.R
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.repositories)
     }
 }
