+++ GitRepositories/app/src/main/res/layout/pull_request_item.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".modules.MainActivity">
 
     <TextView
         android:layout_width="wrap_content"
